Fix resize_images crash and enable it for the correct model input size
resize_images_by_resampling crashed with `Cannot handle this data
type: (1,1), <i8` when resize_images was enabled, because
valid_pxs_img can be int64 and PIL.Image.fromarray rejects that
dtype. The sibling crop_images_to_valid_size already casts to float;
apply the same cast here.

With that fixed, set resize_images: true in
config_diff_img_preprocessing.yaml: with it false, real difference
images stay at final_size (11x11) while the phot-vetting model
expects final_size * upscale_f (55x55).

Verified on the 42-TCE debug subset (job_0-3 from tics_table_debug.csv)
against this commit's own code: 41/42 TCEs produce diff/oot/snr
images at the correct 55x55 shape with no crash; the 1 placeholder
is a TCE with zero valid sectors, unrelated to this fix.

+++ src_preprocessing/diff_img/preprocessing/utils_diff_img.py
@@ -678,7 +678,10 @@ def resize_images_by_resampling(diff_img, oot_img, snr_img, valid_pxs_img, size_
     diff_img = Image.fromarray(diff_img)
     oot_img = Image.fromarray(oot_img)
     snr_img = Image.fromarray(snr_img)
-    valid_pxs_img = Image.fromarray(valid_pxs_img)
+    # valid_pxs_img can be int64, which PIL.Image.fromarray rejects
+    # ("Cannot handle this data type: (1,1), <i8"). Cast to float like the sibling
+    # crop_images_to_valid_size does.
+    valid_pxs_img = Image.fromarray(valid_pxs_img.astype('float'))
 
     diff_img_resize = diff_img.resize(size=(size_w * size_f_w, size_h * size_f_h), resample=Image.Resampling.NEAREST)
     oot_img_resize = oot_img.resize(size=(size_w * size_f_w, size_h * size_f_h), resample=Image.Resampling.NEAREST)
